move out top results indicator
