+++ static/app/components/charts/simpleTableChart.tsx
@@ -12,6 +12,7 @@ import {MetaType} from 'sentry/utils/discover/eventView';
 import {getFieldRenderer} from 'sentry/utils/discover/fieldRenderers';
 import {fieldAlignment} from 'sentry/utils/discover/fields';
 import withOrganization from 'sentry/utils/withOrganization';
+import TopResultsIndicator from 'sentry/views/eventsV2/table/topResultsIndicator';
 import {decodeColumnOrder} from 'sentry/views/eventsV2/utils';
 
 type Props = {
@@ -125,28 +126,6 @@ const HeadCell = styled('div')<HeadCellProps>`
 
 const TableCell = styled('div')`
   padding: ${space(1)} ${space(3)};
-`;
-
-type TopResultsIndicatorProps = {
-  count: number;
-  index: number;
-};
-
-const TopResultsIndicator = styled('div')<TopResultsIndicatorProps>`
-  position: absolute;
-  left: -1px;
-  margin-top: 4.5px;
-  width: 9px;
-  height: 15px;
-  border-radius: 0 3px 3px 0;
-
-  background-color: ${p => {
-    // this background color needs to match the colors used in
-    // app/components/charts/eventsChart so that the ordering matches
-
-    // the color pallete contains n + 2 colors, so we subtract 2 here
-    return p.theme.charts.getColorPalette(p.count - 2)[p.index];
-  }};
 `;
 
 export default withOrganization(SimpleTableChart);
