+++ static/app/views/eventsV2/table/tableView.tsx
@@ -47,6 +47,7 @@ import {getExpandedResults, pushEventViewToLocation} from '../utils';
 import CellAction, {Actions, updateQuery} from './cellAction';
 import ColumnEditModal, {modalCss} from './columnEditModal';
 import TableActions from './tableActions';
+import TopResultsIndicator from './topResultsIndicator';
 import {TableColumn} from './types';
 
 export type TableViewProps = {
@@ -539,28 +540,6 @@ const StyledLink = styled(Link)`
 
 const StyledIcon = styled(IconStack)`
   vertical-align: middle;
-`;
-
-type TopResultsIndicatorProps = {
-  count: number;
-  index: number;
-};
-
-const TopResultsIndicator = styled('div')<TopResultsIndicatorProps>`
-  position: absolute;
-  left: -1px;
-  margin-top: 4.5px;
-  width: 9px;
-  height: 15px;
-  border-radius: 0 3px 3px 0;
-
-  background-color: ${p => {
-    // this background color needs to match the colors used in
-    // app/components/charts/eventsChart so that the ordering matches
-
-    // the color pallete contains n + 2 colors, so we subtract 2 here
-    return p.theme.charts.getColorPalette(p.count - 2)[p.index];
-  }};
 `;
 
 export default withProjects(TableView);
